Раскидал код по разным файлам
