+++ js/create-image.js
@@ -1,0 +1,42 @@
+import {getRandomInteger, getRandomArrayElement} from './util.js'
+import {generateComments} from './generate-comments.js'
+
+const description = [
+  'Пустые шезлонги на пляже',
+  'Указатель пляжа',
+  'Скалы на фоне моря',
+  'Девушка-фотограф на пляже',
+  'Оригинальная подача риса',
+  'Чёрный спорткар',
+  'Поперечный разрез клубники',
+  'Компот из клюквы',
+  'Самолёт, пролетающий над пляжем',
+  'Полка с обувью',
+  'Пляж с кустами',
+  'Моя новая тачка',
+  'Сегодня ем пп еду',
+  'Кто понял, тот понял',
+  'Купил новые тапки',
+  'Лечу на похороны',
+  'Сегодня сходила на хор',
+  'Американская машина в России',
+  'Тапочки для нахождения туалета ночью',
+  'Когда пришёл в парикмахерскую и сказал, чтобы немного подровняли',
+  'Сегодня снова ем пп еду, спустя 2 года',
+  'Последние минуты сегодняшнего солнца',
+  'Краб тоже хочет кушать',
+  'Концерт рамштайна',
+  'На сафари',
+]
+
+const createImage = (i) => {
+  return {
+    id: i,
+    url: `photos/${i}.jpg`,
+    description: getRandomArrayElement(description),
+    likes: getRandomInteger(15, 200),
+    comments: generateComments(getRandomInteger(0, 30)),
+  };
+};
+
+export {createImage}
