+++ js/generate-comment.js
@@ -1,0 +1,30 @@
+import {getRandomInteger} from './util.js'
+
+const names = [
+  'Артем',
+  'Тимур',
+  'Иван',
+  'Софья',
+  'Татьяна',
+  'Олег',
+  'Ольга',
+  'Наталья',
+]
+const commentsMessages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+]
+const generateComment = (i) => {
+  return {
+    id: i,
+    avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+    message: commentsMessages[getRandomInteger(0, commentsMessages.length - 1)],
+    name: names[getRandomInteger(0, names.length - 1)],
+  }
+}
+
+export {generateComment}
