+++ js/generate-comments.js
@@ -1,0 +1,7 @@
+import {generateComment} from './generate-comment.js'
+
+const generateComments = (countComments) => {
+  return Array.from({length: countComments}, ( _, index) => generateComment(index))
+}
+
+export {generateComments}
